res_cdrel_custom: Use double in float getter

+++ res/cdrel_custom/cdrel.h
@@ -180,7 +180,7 @@ struct cdrel_value {
 		int64_t int64;
 		uint64_t uint64;
 		struct timeval tv;
-		float floater;
+		double floater;
 	} values;
 };
